# specs/001-gold-price-app/contracts/market-data.openapi.yaml
openapi: 3.1.0
info:
  title: Gold Price Reporter API
  version: 1.0.0
  description: >
    Public and protected HTTP contracts for the Gold Price Web Application.
    Public routes power the dashboard auto-refresh, historical charts, and
    conversion calculator. Internal routes are protected and reserved for
    scheduler-driven ingestion.
servers:
  - url: http://localhost:3000
paths:
...
  /api/market/convert:
    get:
      summary: Calculate an indicative gold value for a weight and purity
      operationId: getConversionQuote
      parameters:
        - in: query
          name: weight
          required: true
          schema:
            type: number
            exclusiveMinimum: 0
            maximum: 100000
        - in: query
          name: unit
          required: true
          schema:
            type: string
            enum: [toz, g, kg]
        - in: query
          name: purity
          required: true
          schema:
            type: string
            enum: [24k, 22k, 21k, 18k]
        - in: query
          name: currency
          required: true
          schema:
            type: string
            enum: [USD, EGP]
      responses:
        '200':
          description: Conversion quote returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversionResponse'
        '400':
          description: Invalid conversion input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: No trustworthy snapshot available for calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    CronSecret:
      type: apiKey
      in: header
      name: x-cron-secret
  schemas:
    FreshnessState:
      type: string
      enum: [fresh, delayed, stale, unavailable]
    SourceNotice:
      type: object
      required:
        - label
        - sourceName
        - sourceEndpoint
        - lastUpdatedAt
        - freshnessState
        - disclaimer
      properties:
        label:
          type: string
        sourceName:
          type: string
        sourceEndpoint:
          type: string
        lastUpdatedAt:
          type: string
          format: date-time
        freshnessState:
          $ref: '#/components/schemas/FreshnessState'
        disclaimer:
          type: string
...
    ConversionData:
      type: object
      required:
        - requestedWeight
        - unit
        - purity
        - currency
        - spotValue
        - snapshotAt
        - sourceNotice
      properties:
        requestedWeight:
          type: number
        unit:
          type: string
          enum: [toz, g, kg]
        purity:
          type: string
          enum: [24k, 22k, 21k, 18k]
        currency:
          type: string
          enum: [USD, EGP]
        spotValue:
          type: number
        snapshotAt:
          type: string
          format: date-time
        sourceNotice:
          $ref: '#/components/schemas/SourceNotice'
    ConversionResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/ConversionData'
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string